// src/lib/game-data/silk-data/player-data/enum/mapzone-enum-silk.generated.ts
/**
 * This file is auto-generated by a script (See ./scripts/*). Do not edit directly.
 */
// This file is generated by scripts/silk-player-data-gen-types.mts. Do not edit manually.

import * as v from 'valibot';

export type MapZoneSilk =
	| 'NONE' /* 0 */
	| 'TEST_AREA' /* 1 */
	| 'PATH_OF_BONE' /* 2 */
	| 'GREYMOOR' /* 3 */
	| 'SHELLWOOD_THICKET' /* 4 */
	| 'RED_CORAL_GORGE' /* 5 */
	| 'CITY_OF_SONG' /* 6 */
	| 'THE_SLAB' /* 7 */
	| 'GLOOM' /* 8 */
	| 'DUSTPENS' /* 9 */
	| 'BELLTOWN' /* 10 */
	| 'HUNTERS_NEST' /* 11 */
	| 'BONETOWN' /* 12 */
	| 'MOSS_CAVE' /* 13 */
	| 'PHARLOOM_BAY' /* 14 */
	| 'DOCKS' /* 15 */
	| 'WILDS' /* 16 */
	| 'WEAVER_SHRINE' /* 17 */
	| 'BONECHURCH' /* 18 */
	| 'MOSSTOWN' /* 19 */
	| 'LIBRARY' /* 20 */
	| 'CLOVER' /* 21 */
	| 'UNDERSTORE' /* 22 */
	| 'COG_CORE' /* 23 */
	| 'PEAK' /* 24 */
	| 'DUST_MAZE' /* 25 */
	| 'WARD' /* 26 */
	| 'HANG' /* 27 */
	| 'ARBORIUM' /* 28 */
	| 'CRADLE' /* 29 */
	| 'PILGRIMS_REST' /* 30 */
	| 'HALFWAY_HOUSE' /* 31 */
	| 'JUDGE_STEPS' /* 32 */
	| 'MEMORY' /* 33 */
	| 'CRAWLSPACE' /* 34 */
	| 'WISP' /* 35 */
	| 'SWAMP' /* 36 */
	| 'ABYSS' /* 37 */
	| 'AQUEDUCT' /* 38 */
	| 'SURFACE' /* 39 */
	| 'FRONT_GATE' /* 40 */
	| 'CORAL_CAVERNS' /* 41 */;
const MapZoneByName: Record<MapZoneSilk, number> = {
	NONE: 0,
	TEST_AREA: 1,
	PATH_OF_BONE: 2,
	GREYMOOR: 3,
	SHELLWOOD_THICKET: 4,
	RED_CORAL_GORGE: 5,
	CITY_OF_SONG: 6,
	THE_SLAB: 7,
	GLOOM: 8,
	DUSTPENS: 9,
	BELLTOWN: 10,
	HUNTERS_NEST: 11,
	BONETOWN: 12,
	MOSS_CAVE: 13,
	PHARLOOM_BAY: 14,
	DOCKS: 15,
	WILDS: 16,
	WEAVER_SHRINE: 17,
	BONECHURCH: 18,
	MOSSTOWN: 19,
	LIBRARY: 20,
	CLOVER: 21,
	UNDERSTORE: 22,
	COG_CORE: 23,
	PEAK: 24,
	DUST_MAZE: 25,
	WARD: 26,
	HANG: 27,
	ARBORIUM: 28,
	CRADLE: 29,
	PILGRIMS_REST: 30,
	HALFWAY_HOUSE: 31,
	JUDGE_STEPS: 32,
	MEMORY: 33,
	CRAWLSPACE: 34,
	WISP: 35,
	SWAMP: 36,
	ABYSS: 37,
	AQUEDUCT: 38,
	SURFACE: 39,
	FRONT_GATE: 40,
	CORAL_CAVERNS: 41,
};
export const MapZoneSilk = {
	isFlags: false,
	byName: MapZoneByName,
	byId: Object.fromEntries(Object.entries(MapZoneByName).map(([name, id]) => [id, name] as const)) as Record<
		number,
		MapZoneSilk
	>,
	nameList: Object.keys(MapZoneByName) as MapZoneSilk[],
} as const;
export const MapZoneNumberSchemaSilk = v.pipe(v.number(), v.integer(), v.minValue(0), v.maxValue(41));